requirements.txt
decorator==4.0.10
Django==1.10.1
django-braces==1.9.0
django-crispy-forms==1.6.0
django-debug-toolbar==1.6
django-extensions==1.7.4
django-model-utils==2.6
ipython==5.1.0
ipython-genutils==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
pygraphviz==1.3.1
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.1
traitlets==4.3.1
wcwidth==0.1.7
